This upload contains the work of the weekend. Mainly done in xml.c
search.c:
Revised some of the sorting routines.
Removed icomp
Added comparison function compResultsByRank for use with qsort
Added comparison function compResultsByFileNum for use with qsort
Modified sortresultsbyrank and sortresultsbyfilenum to use compResultsByRank
and compResultsByFileNum. The code of these function is simpler and use
less memory.

string.c
Added remove_controls routine to remove control chars

txt.c
Added a call to remove_controls. As stated by Rainer, an external filter can return control chars and should be removed.

index.c
Fixed a segfault when there are files to index but no valid words in none of them. Eg: Index an empty file. It does not have much sense to do things like this, but ...

xml.c
The big changes here. Well, not so much changes but a lot of work done.
Added convertentities.
Properties are stored up to the end tag. Until now, the property will stop when it encounters a '<'. Nested tags are allowed. Eg:

PropertyNames meta1

<meta1>
this
<meta2>
is
</meta2>
the property
</meta1>

Old xml.c stored "this" as property meta1
New xml.c should store "this is the property" as property meta1

NOT yet done in xml.c:

Support for things like:
<meta1 prop1="bla bla" prop2="bla bla">
bla bla
</meta1>

should prop1 and prop2 be indexed or even store as property if they are specified in metanames or propertynames?

New config directive to increase counters as suggested by Bill Moseley