Hi Rainer,
This is what I've had in mind -- and why I've been asking if
gettimeofday() is more precise.

The two "Test" times use gettimeofday().

# Number of hits: 6306
# Search time: 0.130 seconds
# Run time: 0.140 seconds
# Test run time: 0.138
# Test search time: 0.123

I'm not clear if clock() is just returning less precission, or
it's the way we are calling it (my lack of C experience is not helping).
I suspect that it's clock() since CLK_TCK is 100 on my machine
and that is why I'm only seeing two decimal points of data.

I understand that gettimeofday() is not standard on all machines.
./configure can set a define, and then we can call clock() when
gettimeofday() is not supported.

Or we can just show two decimal points -- and accept that many
(if not most) searches will return zero for a search time.  Or we can
not even bother showing the times at all -- but those small search times
are impressive!