Rainer, you will probably have to check my work once again...
This uses gettimeofday, if it's available, otherwise it uses clock().
One thing I'm curious about is why when using clock() I'm getting a negative
zero:

# Number of hits: 134
# Search time: -0.000 seconds
# Run time: 0.010 seconds
# Run: 0.00000000 Search: 0.01000000 End: 0.01000000

That last line are the times just printed.  I'm casing the return form clock() to
a double, could that be causing a problem?

To test you will need to define NO_GETTOD in acconfig.h.