Added -N (Newer) feature to check last modified date
Ranier, Jose, please check my work.

Some of fs.c didn't make any sense to me so I made some changes (which
means I either fixed some things or I broke them)

I changed the orders of the islink checks so that the stat is only done
if FollowSymLinks is NO (the default).  Note that using -l or FollowSymLinks YES
might improve indexing speed a tiny bit.  A few extra stat() calls are saved with
-l (FollowSymLinks YES).

A strace of running indexing is helpful to see these extra calls.

Also, I commented out one call to islink because it was calling it on
dp->d_name which is only true if it happens to be in the current directory.
Later the code correctly pre-pended the directory to dp->d_name and did a lstat
on that path.

I also rearanged the first opendir call because it was opening the directory and reading
the first directory entry even if sw->dirconlist was NULL.  Just didn't make sense that way.

Finally, it would be somewhat smart to cache the stat calls so it's only called once
per file instead of a couple of times (to check for islink, and isdirectory).  But it
probably doesn't matter.