Here's the rewritten indexstring() in index.c.  Please take a look. It was really not much work, and it seems more logical in layout. But please feel free to point out any bad programming.
Since I was there, I added the buzzword feature -- very simple
to add. It's ready to go when the search parser gets rewritten.

I also fixed a few warnings, and used the indent program on a few
more files.

Oh, and the metaname alias feature is in this update.  Rainer's
suggestion on how to specify the aliases in the config file is
probably the way to go, but I didn't have time yet.

Sorry for the huge update -- there aren't that many changes, mostly
reformatting from the indent program.