* Rewrote search parser.  Finally. Well, first edit, anyway.
Can search for ( ) = * \ by preceeding with a backslash.

Not thrilled about the wild card code, but oh well.
I merge the "*" with the preceeding word at the last minute.

Would be better for search.c to leave it as a separate token.
search.c could look at word->next->line to see if it was a "*"
which I think would be cleaner.  Probably should delete all
single "*" chars that are not preceeded by a search word.
Would just need to do a few tricks when printing Parsed Words to keep the
wildcard with the previous word.


Now validates meta names, and metanames can have any chars.


As always, please feel free to offer code imporvements.  Some
swline routines might be better in list.c, perhaps.  Like
inserting a swline into another swline, and removing a node.

I looked carefully at combining index.c code with this code,
but the search parser does a lot of malloc calls.  Probably
not good for indexing speed.  Plus there are a few other
differences in the way the code works between searching and
indexing.


Rainer, you were right.  It took more than a couple of hours -- almost
four.  But I spent an hour trying to fix buzzwords, plus I'm running a
good fever today....