******* Test of Property Limits ***********
Here's the limit code.

I need help with two things (to start off with...)

1) I can't figure out the casts on the qsort compare routines
2) I think the binary search could be cleaned up and validated
for all test cases.  I'm not sure that a binary search is the
best thing for trying to find a range when there may or not
be matches.

I often felt like I was doing things the hard way with some of
code.  If anyone sees some obvious extra work, please let me
know.

The code can easily be pulled back out of swish.  Or it can be clearly
documented that the interface will change.  (Or clearly undocumented ;)

Help is always welcome.