Fixed writing of properties to index file.
This is a partial rewrite of much of docprop.c, but I wanted to check
in before it gets to late in the morning in Europe.

I'm now moving on to the -p property output.  I think I'm going to remove
the array of properties "Prop", and try to tie things together better.

It might be easier to just use a standard -x fmt string when -x is
not used (that is, emulate old ouptput with -x).

I also have a memory leak to fix (need to add a new propValue type
of STRING_FREE (or some such name) so result_output knows to free that string.