This is ONLY a little clean up of merge.c, in an attempt to make it easier to debug.  The code could use some comments.
Jose, there seems to be a problem in the meta merge code that's causing
a segfault.  But that's just a guess.  If I comment out most of the
meta merge code (leaving just the reading of index1, and writing the file and
duplicated properties to the out index) it will not segfault.

I did a CVS update as of June 1, and the segfault is not there, but it is
there sometime before June 14.  My guess is that our attention to cleaning
up memory has made the segfault show up (since it's trying to free something).

For example, in addindexfilelist there's:

    addtofwordtotals(sw->indexlist, sw->Index->filenum, ftotalwords);
    thisFileEntry->docProperties = DupProps(docProperties);

    /* swap meta values for properties */
    swapDocPropertyMetaNames(docProperties, metaFile);

Well, swapDocPropertyMetaNames creates a new list of properties and
returns that (which is ignored above), plus docProperties is
destroyed in that function (which is probably destroyed also in
SwishClose ).

But, commenting that line out is not enough to get rid of the segfault.

Because of the dates, I don't think it's from your change to the new property
structure (although they could be different segfaults).

A note:  I had a segfault yesterday that took me hours to find.  gdb was
reporting the segfault in the same place as this current segfault.
In that case, it turned out that I was writing past the end of an array
and that was somehow messing up other data structures and it was detected
when trying to call free().

You seem to have better luck finding these than I do.