To avoid duplicate code, result_sort.c compFileProps calls docprop.c's Compare_Properties()
Bill, since we have our own qsort, it might be a nice extra feature to be
able to pass a structure to swish_qsort, that then gets passed onto
the compare function.

As is is now, struct file has an extra element 'currentSortProp'
that's just there to pass along which meta entry is currently
being sorted.  Could save a little space by just passing that
right through swish_qsort.