Ok, here's some inital work for writing properties to a separate file, and for reading a file's props all at once, or just a single metaID.
The code's all in #ifdef PROPFILE (see config.h), and the existing
properties have not been removed yet by that #ifdef.

The only way to see it work is to index, and then use -T index_files after
#define PROPFILE is set.

Tomorrow (Thursday) I'll try to get it to sort one property at a time
so Bill can see what differenc that makes.

Maybe that, with Jose's ramdisk that will be enough.  Although it would
be interesting to see what a merge sort would do.