Ok,
So, can you run a few indexing tests, Bill?  the memory should be reduced
for sorting, I believe.

Also, on searches, not a big deal, but memory is freed after each result.

Still need to apply a length limit when sorting the pre-sorted arrays.
The file entry is a bit larger now.

I'd like to hear your results before doing any more work.
MERGE is broken with PROPFILE is defined, no doubt.


Big question, if this stays, is how to deal with the separate .prop file.
Currently is just uses the <indexfilename>.prop path, but that means
the files can't be located on different drives (symlinks?).  But I can't
think of a good way to achive that so -f index1 index2 can find the
right .prop files.