Berkeley DB initial module (db_berkeley_db.c).
To use test it:

1- Download db-3.2.9 from www.sleepycat.com
2- Install it
3- Modify routines initModule_DB and freeModule_DB in db.c. They must called
initModule_DB_db(sw) and freeModule_DB_db(sw) to use db_berkeley_db.c
instead of the db_native.c
4- Add db_berkeley_db.o and -ldb  to the Makefile.
5- PROPFILE must not be defined in config.h (Fix to do)
6- Make and that's all


NOTES:

I do not know very well Berkeley DB. This is my first try with it. So, there
can be much more ways to use it. This is just one and may be improved.

Sendmail and other packages use db files. So be careful if you are using
old versions of db if you install db-3.2.9.