Here's a partial fix for the segfault in removestops.
removestops didn't know that the words and locations are now in a mem_zone,
so the solution is to not efree the memory.

BUT it's still broken with -e (and has been for a while?).  Seems like
we must call unSwapLocData() somewhere when running -e.  Jose, can you take
a quick look.  It's too late for me to think correctly....