I found the segfault in string.c parsetag().
It was stripping trailing blanks/nulls/quotes past the beginning of the string.

Looking at this code makes it obvious why we need an event-based html parser.

For every html file that uses StoreDescription the file is parsed three or FOUR times.
The first is for the title, which only looks a few lines.

Then for StoreDescription HTML <body> parsetag() is called again, but this time it
parses the entire file.

But parsetag() only looks for "<body>" not "<body", so since often <body> conatains
attributes, parsetag() doesn't find the <body> tag.

Then parseHtmlSummary() continues looking, but this time looks for "<body",
and then continues to parse.

Finally, the document is parsed for indexing -- parse number FOUR!
Ok, the title doesn't look too deep unless changed in config.h, which
I have changed to be quite large since so many files contain quite a bit
of header info.

Also note that once the description is found, it's left justified, and newlines
are removed again, and tags are removed.

All this could be done in one pass with an event based parser.

Of course, a new parser would break an undocumented hack in swish where "<" can be
escaped like "\<" instead of &lt;.  Some files I maintain on sunsite actually
have that encoded like that!

libxml2?