New location schema. Uses less memory. Now, during index, locations are colectedin chunks of documents.
It makes heavy use of memzones.

See in config.h:

#define INDEX_DEFAULT_CHUNK_SIZE 10000

This can be adjusted to  whatever you want.

With this value, after proccessing 10000 documents, all locations are
reorganized and coalesced for each entry. This saves about 4 bytes per
location. Also the location_array has been changed for a linked list to
avoid the use of realloc.

Minor changes in mem.c
- Little improvement in Mem_ZoneAlloc for asking just the size we need
(no ROUND_LONG(size)). Also searchs for free size in previous zones
when current becomes full (this is very useful when a Mem_ResetZone
was issued in a memzone).
- Added Mem_ZoneReset. This routine makes a memzone available for reuse.

Yet to do:

- Fix removestops (IgnoreLimit) and merge. You will see warnings when building.
 Ignore them.
- Other bug fix reported by Bill Moseley ("not sksksk" with metanames).

cu
Jose