Fixed some unfreed memory in xml.c
Added MemZones to xml.c html.c ... to use less memory

The trick is as follows. There is a new zone (perDocTmpZone) that
is used to allocate all the temporal data for a document including
locations, title, some temporal buffers, summary, etc...
After the document is proccesed (at the end of do_index_file routine)
this zone is reset (using Mem_ZoneReset) and made it available for the
next document. This saves a lot of emalloc/erealloc/efree operations
and their overhead.