Hi guys,
Here is one more try with the memory issue.

With this fix I am able to index my website (23359 files and 121087 words)
using 50 MB of RAM.

I have found a problem with the new code and large files when being indexed.
The problem was with the memory used to store the temporal locations
(with the new code uses a MemZone). This large files produced a peak
in memory usage inside the MemZone.
Trying to fix it I have added two new routines to get memory for locations
in index.c.

LOCATION *Mem_ZoneNewLocation(MEM_ZONE *head)

LOCATION *Mem_ZoneAddPositionToLocation(void *oldp, MEM_ZONE *head, int frequency)

They seem to avoid the peak.

Also I have added a new routine to mem.c (Mem_ZoneSize). With this routine,
when a file is indexed, I check for the memory in use by the MemZone that
holds the temporal locations and, if it is big enough, I coalesce the
locations.

BTW, -e is also improved in the new code but it has a memory peak usage
at the end.