Added code to grab title (for NoContents), and cleaned up meta/props so that any non-HTML tags are checked for metanames.
Jose:

How hard would it be to have a function I can call to remove a file
from the index.  The problem I'm having is how to do a isoktitle() call
without parsing each document twice.  It would be easier if I could do
it as I'm parsing, and if isoktitle() fails, then abort indexing that file.

That would also be helpful if we decide to parse the docs in chunks, as
I was discussing the other day.