Didn't get to everything on my list today :-(
I added two new config settings:

- ImageLinksMetaName = like HTMLLinksMetaName, it indexes <img src> URLs.
So if you want to find a page with an image of a given name you now can.

- AbsoluteLinks = Use either the <BASE> tag or the file's path to
convert relative URLs to absolute URLs.  Then when you are searching
for images (or links to web pages) you can search (or limit) to
the full path.