Whole bunch of memory leaks in merge and related functions.  Cleaned up some tiny ones too.
I gave up on a few memory leaks as it's late (but they are not big -- like
creating extra indexheader for the three files in merge.

Added a counter to Bill's memory debugger.  That way you can see
which specific allocation was not freed (and then put a check of the counter
and efree(1) to segfault, and gdb backtrace to see exactly where the memory
was allocated.

I can't remember what else.