This is a rather large checkin ;)
Got rid of filearray[], and write all file data
to disk after each file.

To Do:
1) Merge is hosed

2) move -L check to a better place

3) save totalwords per file array, but only save/load if
configured to use totalwords.

4) compress original length in properties

5) separate props and metanames, and only store the properties
needed in the docProperties array.  (currently it is enlarged
each time a prop is added).

6) figure out a better way to save prop position table. Currently it
is a fixed size record

            Prop index 0      Prop index 1
          <seekpos><length>:<seekpos><length>:...

So, to lookup a property it's just a matter of seeking directly to
the table, reading two longs, then seeking in the prop file.

And it's based on the number of PropertyNames.  So

   PropertyNames propA propB propC

for 1,000,000 files add 24M to the index file. 8 bytes x 3 extra props.

The reason for this is because Ask was indexing 2.4M files, and a search
would create an array big enough to hold all those entires.  Very slow
to start up.  Plus every word found (not just a result) would cause
the 40 byte struct file to be loaded.  The property pointers only need
to be read when printing results, not before.

I can't imaging there's not a few bugs considering the amount of code
changed.