Several modifications
- getrank moved from search.c to rank.c (see previous update)
- Better performance for sorting results (this includes sorting by rank).
To do that I have removed some unneeded repetitive code in sortresults and
I have also applied MemZones to this part of the code to allow the
malloc/free storming that slows down the search proccess when thousands
of hits are returned (Ask's problem?).
- Computing of ranking is delayed when possible. So, if you sort for
something different than rank, only rank is computed for the results
specified in maxhits. This delay is not possible when "and", "or" "phrase"
operation are being used. To do that, I asign the value of -1 to rank that
means "not yet ranked". This also has a lack that I think it is not important:
rank is not normalized to 10000 when sorting from something different from
rank; well, I think that if we do not sort by rank, rank is not the most
important parameter.