Jose fixed this bug where -w 'i*' found many words and used lot's of memory.
From Jose:

"Well, it was a real memory problem. The problem showed when switching
to MemZones. The guilty routine was mergeresulthashlist that was
allocating too much memory space. It was allocating memory for merging
pairs of results. This should be a CPU bottleneck even with malloc/free.
Now, instead of allocating memory for pair of results, it allocates
and merge the whole bunch of results that shares the same filenum.
So, we still can use MemZones but the problem seems gone."