I'm finally checking in an old update to provide the total files in the swish library.
I added code that is *not* thread-safe in swish2.c.  Can someone look and see if
there's a better way?  I couldn't see a way to allocate memory without having
a memory leak.

I suppose I could allocate the memory, and then document that the caller must free the
the memory with efree.  Is that the better way to go?

Also added a test for the perl routing, and also minor docs update.