Hi guys,
This upload fixes many things in the "updateable" index database.

The current state of art is:

- It is working for adding files (I think so).

But it has some lacks that need to be fixed:

- File space is not reused when a page is empty after an update. This
is specially shown in the worddata file (wdat extension).

- If a file already exists in the database, the a dup entry is produced.

An there is also a to do section:

- Add a way to delete files (this will allow update). As discussed this
needs an extra index in some way to handle the relation (filename, filenum).
This option will also alow updates (delete + insert).


BTW, if you want to try, just uncomment USE_BTREE in config.h.
To update an index just add a -u in the command line. Eg:

swish-e -c file.conf -u

or

swish-e -i myfile -u -f indexfile

If you specify -u, the code takes care for an existing index file. The
original header values are more relevant than the possible new ones in
the config file. In other words, when updating the index file, the sequence
for the header options is:
1- Default values
2- Config file
3- Header values
(3 overwrites 2 , 2 overwrites 1)

Any comments are welcome. Bugs are also welcome :(

cu
jose