Fix for the merge crash.
The problem with merge was related with the new filenums asigned to
worddata while merging because they were not consecutive (in the index
proccess are consecutive).
This means that we can not use coalesce_word_locations routine to save
memory (this routine also swaps location data to files when using -e).

PROS: None :(
CONS: Unfortunately -e is useless in merge :(