Allow finer tuning to the page display.
In response to Greg Fenton's suggested features to swish.cgi.

He had suggested ways to customize the page number navigation
display, a feature to use different colors for each result,
and percentage and "stars" to replace the ranking number.

I only implemented the page number - but only in a way that makes
the raw data available to the templates.  See swish.tt for an
example.

The others can be implemented directly in the search.tt template
or in a combination of the swish.tmpl and the TemplateHTMLTemplate.pm
file.  I want keep the display code in the display module or template.