Add refcount checking in the SWISH::API Perl module. That means the Perl API works like you would expect in perl in that parent objects are not destroyed until child objects are destroyted.
It also lets people eat up memory:

sub first_hit {
  my $handle = SWISH::API->new;
  my $results = $handle->Query( "foo" );
  my $first_hit = $results->NextResult;
  return $first_hit;
}

my $first_one = first_hit();

my $path = $first_one->Property('swishdocpath');

Now since $first_one is a SWISH::Result object, the results list
is still in memory (so if a search for "foo" found 100,000 results, those
are all still in memory), and the index is still opened.
As soon as $first_one goes out of scope, or is undefined, the results will
be freed and the index closed.

I suppose that should be documented.