Enabled AM_MAINTAINER_MODE
This prevents the Makefile from updating the build tools.
The hope is that it will allow normal users to build from cvs
with less trouble.

Developers with a current set of autotools (autoconf, automake, autoheader,
libtool) can run ./configure --enable-maintainer-mode to re-enable the rules.
Or I think we can just try and remember to run ./bootstrap when any change
is made where the tools need to be run -- such as modifying configure.in
or any Makefile.AM.

http://sources.redhat.com/automake/automake.html#maintainer-mode