Here's that patch recommended the other day.  I'm not completely comfortable with the change -- it was not right before but it still doesn't work right. For one thing file names with tabs or other chars < a space are replaced by a single space -- so a filename with a tab will still be wrong.
I could be convinced to remove all checks so that any file name would be
indexed (^A^B\t^C could be a valid file name).  But clearly a \n in file
names would cause problems when parsing output from swish-e.

The other reason I had that code in there was to compress lots of white
space --  mostly to save space and to reduce the number of chars that must
be processed when highlighting.

The best solution would probably be to accecpt all chars and then filter
when printing with the default output mode.

Anyone have opinions?