Move from int to unsigned int in posdata array trying to fix Steve's big file problem. At least, in my linux PC it seems to work with a file with 0xffffff words.
Also added a new conf feature: CompressPositions yes|no
When set to yes it compress word's position array with zlib (as it has been done). Although zlib is pretty fast uncompressing, the use of it can affect search performance. When we search for a word there is not any problem, only 1 uncompress is made. But when we search for a*, and the search returns 3000 words, there are 3000 calls to the zlib uncompressing routine and this can take long time...

BTW, I will be out for the next 4 days because of vacation.