This fix improves compress_worddata routine.
Symptoms: Some users indexing in -e mode are experimenting annoying messages
of "Out of RAM" when entering this routine.

I think that the reason is that in order to call zlib's compress2 routine
an extra buffer is allocated to store the compressed result and this buffer
can be very big for very common words (a, of, the, etc).

Fortunately, we can use zlib's low level routines (deflateInit, deflate and
deflateEnd) that can compress in chunks. The trick is storing the compressed
output in the original buffer taking care of buffer overrun.