Added -R command line switch to allow toggling the ranking scheme. This preserves the current ranking model (the default) and allows experimentation/testing with the new IDF model.
Requires that IgnoreTotalWordCountWhenRanking be set to
'no' (false). Added an additional field to the index to store
the total word position count for each index. This will allow
for relative density calculation in addition to IDF.

I added code in rank.c to allow for future ranking development
by adding routines that follow different ranking models. This will
allow a user to switch ranking schemes on the fly to help tune
performance.

TODO:

* the adjustWordPositions() function needs to account for
the addition of the total_word_positions count in the index. I
ran out of steam in trying to decipher how that code works -- Jose
is just a damn genius.

* add API to SetRankScheme for Perl and C. I need to ask Bill how
to do that.