Hi Peter,
The numbers has no special meaning. As you say this is to prevent reading the wrong files. So, this value means: "This is a swish file".

How it is calculated? here is the big secret: I was born on 21-July-1963 (21-07-1963) = 210763 ;)
The other 22 has no special meaning. It is for padding the long number.

It is still used in db_native.c (DB_CheckHeader routine).

Change them if you want.

cu
Jose

Peter Karman escribió:

> I see this comment in swish.h:
>
> /* $$$ THESE NEED TO BE UPGRADED WHEN THE INDEX FORMAT CHANGES */
>
> #ifdef USE_BTREE
> #define SWISH_MAGIC 11076322L
> #else
> #define SWISH_MAGIC 21076322L
> #endif
>
>
> and I wonder if my recent change to the header to track total word positions means I need to change these values. Is this SWISH_MAGIC still used? If so, how are the values calculated?
>
> I assume this is to help prevent swish-e from trying to read indexes created with an older version of swish-e. There's an item on the to-do list about this too.
>
> Just looking for a little history...
>
> thanks.