Here's the "fix" for searching for '(foo)' including the parens.
Peter, note that you can search using backslashes, just not as the first char:

   -w 'x\(foo\)'

works.  There's a comment in swish_words.c:

/* Can't backslash here -- (because this code should really be include in swish_words.c) */

Which makes no sense to me... ;)  Actually, the problem may not be in that bit of code,
but how next_token() splits up the text.

Wnat to spend some time with gdb?