Seem I forot to do a cvs ci.
** Question:

I changed the ParserWarnLevel default from 0 to 2 a while back.

I like having a separate setting for libxml2 (instead of using, say, -v setting),
but I wonder if -v0 might be special.

The reason I ask is so there's a way to disable warnings when not using a config file.

Or maybe we could add a switch for setting the parser warn level from the command line?