Here's a few updates to what you did.
BTW:

dnl It would be better not to rely on a specific html file but how else to do it?
dnl We can 'find' but is that a standard command?
dnl if test -e "html/api.html"; then
somehtml=`find html -name '*.html' -print`

find is probably everywhere, but proabably not considered portable.
I would think you could check return valudes, too.

But where's it looking?  Relative to the current directory -- so
there won't be any html if not building in the source tree, right?

Why not use tes to look for a file?  See below for details.